Implemented a gradient descent algorithm to try and find best fits for x,y coordinates of a cohort of forest cover data points with fire, water and road points.
The algorithm seems to have something wrong with it at the moment since
the cost function starts out decreasing, but at near a consistent value
it starts increasing again rather than limiting out.  This effect seems
to happen regardless of what value I use for alpha (as long as I also
use sufficient iterations to allow effect to be reached).