Mapping from fire/water/road distances to x, y coordinates seems to be working reasonably well with random fake data.  Sometimes it seemed to get stuck in a local minima, and when I looked in both cases there were a small number of points that had significantly more error than the others.  I manually changed the hypothesis for those points (just to give it a try, assigning them the negative of their coordinates) and that seemed to work well to get it on track for a better minimum.
Of course this is all still relying on the assumption that road and
water are treated as points rather than shapes (i.e. if points are
nearish to each other and the closest hydrology is a lake they will both
be measuring to the same point on the lake rather than just whatever is
the closest lake point to them).  Still I suspect that even if these
assumptions are not valid, it may still may be okay as long as the
points are in a reasonably close neighborhood.