recentering with rotation to match up hypothesized points with true points.  Seems to work pretty well - though it appears there may be trouble when two of the fixed points are very near to each other.  With one example that turned out that way there were 2 or 3 hypothesized points that had relatively small error but were a fair distance from their true counterparts.  On creating a new example with fixed points not too close to each other this was not a problem.  In both cases the hypothesized fixed points seemed to match up quite well with the true fixed points.
More studying may be called for wrt pernicious cases as well as further
testing to verify that it does consistently work.

Sometimes there are points that seem to get caught in a sub-optimal
minima.  One technique that might be worth trying more is to recenter
the points on a fixed point that the problem point is far off the true
distance, then replacing the hypothesized problem point with its
negative.  I've only tried this once so far, but it worked well in that
case.