Begin implementing option for normalizing distance features in gradient descent.  As yet untested other than to check that the default choice (of no normalization) still works as before.
Other (important) change - fixed partial derivative of cost function
with respect to fixed point variables.  I had not been dividing those
partials by 2m as I had done with the other partials.  Adding that
division in seems to have stabilized the convergence so that I can
choose a larger learning rate (alpha) and attain similar results in much
fewer iterations (in tests it often seemed to be 10x or more improvement
in the required iterations).