Began work on creating cohorts from data set.  Next will need to run the gradient descent on the cohorts and then match cohorts back up into larger set.
Did a little bit of testing of correct convergence for small data sets.
For sets of 2 and 3 the algorithm sometimes converged to the correct
solution, but often enough did not.  With 4 points there was one strange
example where the algorithm failed to converge - for some reason the
jiggle function did not seem to be sufficient.  Several other 4 point
sets converged correctly.  I am going to go with the idea that 4 points
will in general be sufficient.