Automating cohort alignment.  Working pretty well, but fixed point alignment seems to not be working correctly.
Should investigate optimizing the jiggle phase of the optimization
process.  It is mostly working fine, but when there are a lot of points
in the cohort it often takes a lot of iterations to get to minimum.
Currently I have it set to jiggle just one point at a time because with
smaller sets of points, when I was jiggling more points at a time it
sometimes jumped back and forth between local minima.  Might work to
jiggle larger number of points at a time grouped by what fix point to
reflect across as well as (possibly) how far from true they are relative
to the worst points.