Issue with aligning cohorts now seems to be resolved.
Proceeding I still need to put in some code to see if fixed points are
likely the same and only include them in the full list of they are new.
Also need to make cohort creation overlap cohorts more robustly (so
don't unnecessarily end up with cohorts that don't contribute because
they don't overlap enough/at all with any of the other cohorts.)

Might already be at the point though where it would be worthwile to test
a little with real data.