geographic_test:  modified distance calculation function for finding cohorts - vectorized the operations and added in a filter for first comparing vertical distance to water - elevation between the locus point and potential cohort points (should be approx. the same if they are relative to the same water point source) Also added ability to easily change the radius used to find cohort of points.
x,y clustering.ipynb:
started as a copy of different python notebook.  Using it to explore the
process using the actual training data.

So far adjusting the cohort radius and adding the hydrology elevation
filter seem promising (and improve the speed of the cohort finding
part).  The gradient descent part on the cohorts does not seem to be
doing well now though so that is probably the next thing to work on.
Some of the cohorts seemed stuck at high cost, others got stuck (or
alternated) at lower costs.  Maybe not possible to get cost too close to
0 for the real world data (did get a couple of cohorts with costs around
10 which seems like it might be pretty reasonable).