Some minor changes and some testing of new techniques.
Changes include copying plot3d to geographic_test for plotting of 3d
values after x,y mapping has been completed.

New tests in the python notebook include removing problem points rather
than trying to jiggle them.  For the initial test this seemed to work
well enough so next step is to implement it as part of the automated
process (may want to leave jiggling as an option).  Another change that
should likely be made is to adjust the cost threshold from .5 to
something more around the size of 10.