Begin work on progressive process of building model.  Start with a cohort then build out progressively from there to ensure that cohorts overlap.
This isn't currently performing very well.  It hasn't done a good job of getting broad cohort coverage.  From visual inspection of the results it seems to be sometimes fitting the data poorly (e.g. in one set, ended up with a bunch of points in close x, y proximity but different elevations which would suggest gathering many data points on a cliff which seems unlikely).  Another issue is that the fixed point tolerance may need to be adjusted as at the end of the process it sometimes ends up with a lot of fixed points fairly close to each other rather than treating them as the same point (which may or may not be correct - maybe there are multiple fire points in close proximity).

Another issue that needs to be addressed is that the examine_results function needs to be adjusted to consider fixed points that aren't just the initial ones in the data structure.