Start of new approach to fitting points and fixed points: First get a set (fire, water, road) of fixed points that seems to be good (at this point, just using the cohort fitting method I've been using until it finds a cohort with cost below the threshold). Second, run gradient descent on the broader data set against the presumed good fixed points.  This in theory will get all the data points that had those as their closest fixed points (and possibly some additional points that incidentally fit).  To get the well fitting points, use GeoParser.examine_results function and filter out points with total error above some threshold value. That's as far as I've gotten at this point, but in future might want to expand out to run on different sets of fixed points and then combine the results somehow. First attempt with test data seemed to work quite well, but took long time to run bgfs/gradient descent to fit the points to the fixed points (this was using a 300 point data set).  Might try running on smaller sets of data at a time.
I didn't time the first trial, it had found fixed points with an initial fitting cost much less than 1. There were ~220 points that fit well accordint to
the results, and the plot lined up quite nicely.
The second trial took 20 minutes to fit the 300 points, with 352 iterations.
The fixed points came from an initial fitting with cost around 20.  After
fitting the data to the fixed points, the smallest error was ~5 and quickly
went up from there.  In other words, none of the points fit very well.  It
appears from this that it might be quite important to get a pretty good initial
fit for the fixed points.