Refactored a bunch.  Main goal of the refactoring was to break out different parsing approaches into their own classes and have a base class that just did the basic/shared parts of the process.
Now have progressive parser which attempts to find one good fitting cohort
and build out from there.  Doesn't work super well with most of my tests
but does work some.  I suspect it might work better on the real data which
is more densely populated.

The fpblanket parser looks for well fitting cohorts (and keeps the number
of points examined while doing so small to maximize likelihood of good
fit).  Takes the fixed points from well fitted cohorts and looks for all
other points that fit those fixed points well.  This seems to work quite
well but can take a long time to attempt to fit against the rest of the points.
Next step is likely to try and reduce the number of times that step is run
and possibly to only attempt to fit against a subset rather than the full set
of points.  In the long run that will probably be necessary as the real data
set will be much larger.  One possibility is to try to fit individual or small groups of points at a time.  Possibly more efficient than one large set?